+++ src/main/java/net/forcemaster_rpg/item/weapons/WeaponsRegister.java
@@ -86,48 +86,48 @@ public class WeaponsRegister {
 
     public static final Weapon.Entry wooden_knuckle = knuckles("wooden_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.WOOD, () -> Ingredient.fromTag(ItemTags.PLANKS)), 1.5F, false)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 0.5F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 2.0F))
             ;
     public static final Weapon.Entry stone_knuckle = knuckles("stone_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.STONE, () -> Ingredient.fromTag(ItemTags.STONE_TOOL_MATERIALS)), 2.5F, false)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 1.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 2.5F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.01F))
             ;
     public static final Weapon.Entry iron_knuckle = knuckles("iron_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.IRON, () -> Ingredient.ofItems(Items.IRON_INGOT)), 3.5F, false)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 2.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 3.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.02F))
             ;
     public static final Weapon.Entry golden_knuckle = knuckles("golden_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.GOLD, () -> Ingredient.ofItems(Items.GOLD_INGOT)), 1.5F, false)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 2.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 2.5F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.02F))
             ;
     public static final Weapon.Entry diamond_knuckle = knuckles("diamond_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.DIAMOND, () -> Ingredient.ofItems(Items.DIAMOND)), 4.5F, false)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 3.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 4.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.03F))
             ;
     public static final Weapon.Entry netherite_knuckle = knuckles("netherite_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.NETHERITE_INGOT)), 5.5F, true)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 4.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 5.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.05F))
             ;
     public static final Weapon.Entry legendary_golden_knuckle = knuckles("legendary_golden_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.DIAMOND, () -> Ingredient.ofItems(Items.GOLD_INGOT)),4.5F, true)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 4.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 5.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.05F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.CRITICAL_CHANCE.id, 0.05F))
             ;
     public static final Weapon.Entry guardian_knuckle = knuckles("guardian_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.DIAMOND, () -> Ingredient.ofItems(Items.IRON_INGOT)),5.0F, true)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 3.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 4.0F))
             .attribute(armorAddition(4.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.025F))
             ;
     public static final Weapon.Entry bloody_knuckle = knuckles("bloody_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.NETHERITE_INGOT)),6.5F, true)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 5.0F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 6.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.07F))
             .attribute(ItemConfig.Attribute.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:lifesteal_modifier")),0.10F))
             .attribute(ItemConfig.Attribute.multiply(Objects.requireNonNull(Identifier.tryParse("more_rpg_classes:arcane_fuse_modifier")),0.05F))